# student-management-service/swagger.yaml
openapi: 3.0.0
info:
  title: Student Management Service API
  version: 1.0.0
  description: API for managing student enrollments, payments, certificates, courses, quizzes, and more
paths:
...
  /certificates:
    post:
      summary: Generate a new certificate
      tags:
        - Certificates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enrollmentId:
                  type: string
      responses:
        '201':
          description: Certificate generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  certificate:
                    type: object
                    properties:
                      id:
                        type: string
                      enrollmentId:
                        type: string
                      issuedAt:
                        type: string
                        format: date-time
                      pdfUrl:
                        type: string
                      verificationCode:
                        type: string
                      studentName:
                        type: string
                      courseTitle:
                        type: string
        '400':
          description: Bad request
        '404':
          description: Enrollment, user, or course not found

    get:
      summary: List certificates
      tags:
        - Certificates
      parameters:
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of certificates
          content:
            application/json:
              schema:
                type: object
                properties:
                  certificates:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        issuedAt:
                          type: string
                          format: date-time
                        pdfUrl:
                          type: string
                        verificationCode:
                          type: string
                        studentName:
                          type: string
                        courseTitle:
                          type: string
...
  /courses:
    get:
      summary: List courses
      tags:
        - Courses
      parameters:
        - in: header
          name: Accept-Language
          schema:
            type: string
          required: false
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
  /courses/{id}:
    get:
      summary: Get course details
      tags:
        - Courses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: Accept-Language
          schema:
            type: string
          required: false
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
...
  /student-assignments:
    get:
      summary: List student assignments
      tags:
        - Student Assignments
      security:
        - CognitoUserPools: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    post:
      summary: Create a new student assignment
      tags:
        - Student Assignments
      security:
        - CognitoUserPools: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
  /student-assignments/{id}:
    get:
      summary: Get student assignment details
      tags:
        - Student Assignments
      security:
        - CognitoUserPools: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    put:
      summary: Update a student assignment
      tags:
        - Student Assignments
      security:
        - CognitoUserPools: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: Delete a student assignment
      tags:
        - Student Assignments
      security:
        - CognitoUserPools: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
...
  /assignments/{id}:
    get:
      summary: Get assignment details
      tags:
        - Assignments
      security:
        - CognitoUserPools: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    put:
      summary: Update an assignment
      tags:
        - Assignments
      security:
        - CognitoUserPools: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: Delete an assignment
      tags:
        - Assignments
      security:
        - CognitoUserPools: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    CognitoUserPools:
      type: apiKey
      name: Authorization
      in: header